+++ src/functions/variance.hpp
@@ -1,3 +1,0 @@
-#pragma once
-extern void scale(PlayerObject* player, bool isMode);
-extern void wide(PlayerObject* player, bool isMode);