+++ src/util.hpp
@@ -1,3 +1,0 @@
-#pragma once
-extern bool isValid(PlayerObject* player, bool detectPlayer);
-extern int playerIndex(PlayerObject* player);